refact: BasicAuth structをconfigから移動した
