+++ backend/services/admin.go
@@ -1,9 +1,9 @@
 package services
 
 import (
-	"backend/config"
 	"backend/models"
 	"backend/stores"
+	"backend/utils"
 
 	"golang.org/x/crypto/bcrypt"
 )
@@ -12,7 +12,7 @@ type (
 	AdminService interface {
 		GetAdminById(id string) (*models.Admin, error)
 		GetAdminByUsername(username string) (*models.Admin, error)
-		CreateAdmin(creds *config.BasicAuth) (string, error)
+		CreateAdmin(creds *utils.BasicAuth) (string, error)
 		// UpdateAdminById(user *models.User) (string, error)
 		// DeleteAdmin(id string) error
 	}
@@ -22,7 +22,7 @@ type (
 	}
 )
 
-func (s *adminService) CreateAdmin(creds *config.BasicAuth) (string, error) {
+func (s *adminService) CreateAdmin(creds *utils.BasicAuth) (string, error) {
 	encryptedPassword, err := bcrypt.GenerateFromPassword(
 		[]byte(creds.Password),
 		bcrypt.DefaultCost,
